Consolidate CES connection ownership into ces-runtime singleton
Move the full CES startup orchestration (process start, handshake,
reconnect wiring, client-ref tracking) into the ces-runtime singleton
so lifecycle.ts no longer touches secure-keys or the CES internals.
Routes import getCesClient directly from ces-runtime instead of pulling
it through the secrets-deps DI shim.
