+++ assistant/src/__tests__/secret-routes-platform-proxy.test.ts
@@ -178,7 +178,6 @@ describe("secret routes managed proxy registry sync", () => {
     platformBaseUrlOverride = undefined;
     providerRefreshCalls = 0;
     registerSecretsDeps({
-      getCesClient: () => undefined,
       onProviderCredentialsChanged: () => {
         providerRefreshCalls++;
       },
