+++ assistant/src/runtime/routes/secrets-deps.ts
@@ -1,15 +1,12 @@
 /**
  * Module-level singleton for secrets route dependencies.
  *
- * The daemon server registers its CES client accessor and provider-reload
- * callback at startup via {@link registerSecretsDeps}. Route handlers import
- * {@link getSecretsDeps} to access them without DI.
+ * The daemon registers its provider-reload callback at startup via
+ * {@link registerSecretsDeps}. Route handlers import {@link getSecretsDeps} to
+ * access it without DI.
  */
 
-import type { CesClient } from "../../credential-execution/client.js";
-
 export interface SecretsDeps {
-  getCesClient: () => CesClient | undefined;
   onProviderCredentialsChanged: () => void | Promise<void>;
 }
